backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==1.2.0
httpx==0.24.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
slowapi==0.1.9
redis==4.5.5
